Made improvements on movement and rotation
read the TODO's
player rotates, will move based on direction of rotation, needs some
changes still but is working for the most part, need the touch check so
it doesn't permanently rotate.

could maybe stop it rotating while it's moving and then have it rotate
again once it stops moving, just need to get it to stop properly.